+++ webservice/src/main/resources/assets/scripts/modules/Modal.js
@@ -155,12 +155,14 @@ define(function(require) {
 
           self.trigger('closeModal');
         });
-        
+      }
+
+      if(this.options.escape) {
         $(document).one('keyup.dismiss.modal', function (e) {
           e.which == 27 && self.trigger('closeModal');
 
           if (self.options.content && self.options.content.trigger) {
-            e.which == 27 && self.options.content.trigger('shown', self);
+            e.which == 27 && self.options.content.trigger('closeModal', self);
           }
         });
       }